fix gp test
